requirements/production.txt
Django
psycopg2
South
factory-boy
Unipath
gunicorn
django-braces
django-spurl
django-skd-tools==0.0.6
django-ckeditor
